backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
psycopg2
django-filter==24.3
requests
django-extensions



